Added additional purchasable items to the store

+++ Assets/Scripts/MoneySystem.cs
@@ -5,7 +5,7 @@ using UnityEngine.UI;
 public class MoneySystem : MonoBehaviour
 {
     [SerializeField] public TextMeshProUGUI moneyText;
-    static public int playerMoney = 10;
+    static public double playerMoney = 10;
     static public MoneySystem instance;
 
     void Awake()
